refactor: engine part codes

+++ Ch07_SceneRendering/app/src/main/cpp/Engine/Rendering/Mesh/IMesh.h
@@ -11,6 +11,6 @@ public:
     virtual const std::vector<Vertex>& GetVerticesWithNormals() const = 0;
     virtual const std::vector<uint32_t>& GetIndices() const = 0;
     
-    virtual uint32_t GetVertexCount() const = 0;
-    virtual uint32_t GetIndexCount() const = 0;
+    virtual uint64_t GetVertexCount() const = 0;
+    virtual uint64_t GetIndexCount() const = 0;
 };
